Integer to Long

+++ nwt-cinema-movies/src/main/java/ba/unsa/etf/nwtcinemamovies/interfaces/MovieRepository.java
@@ -7,6 +7,6 @@ import ba.unsa.etf.nwtcinemamovies.models.Movie;
  *
  * @author eceric
  */
-public interface MovieRepository<M, I extends Number> extends AbstractRepository<Movie, Integer> {
+public interface MovieRepository<M, I extends Number> extends AbstractRepository<Movie, Long> {
 
 }
