+++ nwt-cinema-movies/src/main/java/ba/unsa/etf/nwtcinemamovies/interfaces/UserRoleRepository.java
@@ -7,6 +7,6 @@ import ba.unsa.etf.nwtcinemamovies.models.UserRole;
  *
  * @author eceric
  */
-public interface UserRoleRepository<U, I extends Number> extends AbstractRepository<UserRole, Integer> {
+public interface UserRoleRepository<U, I extends Number> extends AbstractRepository<UserRole, Long> {
 
 }
